+++ CHANGELOG.md
@@ -11,6 +11,12 @@ All notable changes to this project will be documented in this file. The format 
 - Fixed:
   - Update Listings component to sort listings by status ([#1585](https://github.com/bloom-housing/bloom/pull/1585))
 
+
+### UI Components
+
+- Added:
+  - Add ResponsiveTable for pricing
+
 ## v1.0.5 08/03/2021
 
 - Added: